+++ confluence/embedded-confluence-public/src/edit-page/__tests__/EditPage.test.tsx
@@ -35,7 +35,7 @@ it('should use default locale to localize text by if no locale is provided', asy
 
   expect(await findByTestId('confluence-page-iframe')).toHaveAttribute(
     'src',
-    `${mockDefaultProtocol}//${mockDefaultHost}/wiki/spaces/TEST/pages/edit-embed/123?parentProduct=test&parentProductContentContainerId=10000&locale=en_US`,
+    `${mockDefaultProtocol}//${mockDefaultHost}/wiki/spaces/TEST/pages/edit-embed/123?parentProduct=test&parentProductContentContainerId=10000&locale=en-US`,
   );
 });
 
@@ -46,7 +46,7 @@ it('should localize text by locale provided by React prop', async () => {
 
   expect(await findByTestId('confluence-page-iframe')).toHaveAttribute(
     'src',
-    `${mockDefaultProtocol}//${mockDefaultHost}/wiki/spaces/TEST/pages/edit-embed/123?parentProduct=test&parentProductContentContainerId=10000&locale=zh_CN`,
+    `${mockDefaultProtocol}//${mockDefaultHost}/wiki/spaces/TEST/pages/edit-embed/123?parentProduct=test&parentProductContentContainerId=10000&locale=zh-CN`,
   );
 });
 
@@ -59,6 +59,6 @@ it('should localize text by locale provided by IntlProvider', async () => {
 
   expect(await findByTestId('confluence-page-iframe')).toHaveAttribute(
     'src',
-    `${mockDefaultProtocol}//${mockDefaultHost}/wiki/spaces/TEST/pages/edit-embed/123?parentProduct=test&parentProductContentContainerId=10000&locale=zh_CN`,
+    `${mockDefaultProtocol}//${mockDefaultHost}/wiki/spaces/TEST/pages/edit-embed/123?parentProduct=test&parentProductContentContainerId=10000&locale=zh-CN`,
   );
 });