+++ linking-platform/linking-common/src/promiseDebounce.ts
@@ -12,21 +12,29 @@ export function promiseDebounce<
   Args extends unknown[],
   ResolveType extends unknown
 >(cb: (...args: Args) => Promise<ResolveType>, time: number) {
-  // Setup array of resolves to promise a result to
-  const resolvers: Array<(res: ResolveType) => void> = [];
+  // Setup array of resolves/rejects to promise a result to
+  let resolvers: Array<(res: ResolveType) => void> = [];
+  let rejectors: Array<(reason: any) => void> = [];
 
   // The function to be debounced and only run after x interval
   // Once called await the promise that will yield our results and
   // resolve all resolvers with the value
   const fn = async (...args: Args) => {
-    const result = await cb(...args);
+    try {
+      const result = await cb(...args);
 
-    while (resolvers.length) {
-      const resolve = resolvers.pop();
-      if (resolve) {
+      for (const resolve of resolvers) {
         resolve(result);
       }
+    } catch (err: unknown) {
+      for (const reject of rejectors) {
+        reject(err);
+      }
     }
+
+    // clean up arrays after execution
+    resolvers = [];
+    rejectors = [];
   };
 
   const debounced = debounce(fn, time);
@@ -34,8 +42,9 @@ export function promiseDebounce<
   // Returns a promise that fires the debounce fn and pushes a resolver
   // to have the debounced result resolved
   return (...args: Args) =>
-    new Promise<ResolveType>(resolve => {
+    new Promise<ResolveType>((resolve, reject) => {
       resolvers.push(resolve);
+      rejectors.push(reject);
       debounced(...args);
     });
 }