+++ media/smart-card/examples/13-flexible-ui.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-import FlexibleUiExample from './flexible-ui';
-
-export default () => <FlexibleUiExample />;